Revert "make xtra download ioctl complete without error"
This reverts commit 2b4c29dcf42bb8c7fb2550b9d6e860563be5b8e1.
On closer inspection this "fix" didn't do what I thought it did. The first ioctl actually timed out waiting for a reply (because it obviously doesn't send one). The rest appeared to succeed and quickly drove the timed out first off the top of my screen. They were actually failing too but the return code of the function without wait for reply doesn't return the same values as the one that does, which I didn't realise, so the error wasn't caught by my code.